Fix missing mkdir command
https://github.com/xtormentiumx pointed out a bug in the documentation, where the Docker Compose will fail to bring up the containers.

This is because of a missing bind directory, as we were not creating it. I had that directory created while testing and documenting this and missed that I had initially had to create it. This commit simply adds a mkdir -p data/backend and improves the language a bit to explain what it's doing.